backend/requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5
flake8==3.8.3
pytest==5.4.3
pytest-cov==2.9.0
pytest-django==3.9.0
django-filter==2.3.0
djangorestframework==3.11.0
django-rest-knox==4.1.0
django-cors-headers==3.4.0
django-extensions==2.2.9
robotframework-requests==0.7.2
